LPS-86550 Match class name
